+++ core/api_container/main.go
@@ -194,7 +194,6 @@ func createServiceNetworkAndLambdaStore(
 		freeIpAddrTracker,
 		args.ShouldPublishPorts,
 		filesArtifactExpander,
-		dockerNetworkId,
 		enclaveId,
 	)
 
@@ -208,6 +207,7 @@ func createServiceNetworkAndLambdaStore(
 		isPartitioningEnabled,
 		freeIpAddrTracker,
 		dockerManager,
+		dockerNetworkId,
 		enclaveDataVol,
 		userServiceLauncher,
 		networkingSidecarManager)