+++ core/api_container/server/service_network/service_network_impl.go
@@ -22,6 +22,7 @@ import (
 	"io"
 	"net"
 	"os"
+	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -37,7 +38,8 @@ const (
 
 // Information that gets created with a service's registration
 type serviceRegistrationInfo struct {
-	ipAddr net.IP
+	serviceGUID      service_network_types.ServiceGUID
+	ipAddr           net.IP
 	serviceDirectory *enclave_data_volume.ServiceDirectory
 }
 
@@ -53,14 +55,14 @@ type serviceRunInfo struct {
 /*
 This is the in-memory representation of the service network that the API container will manipulate. To make
 	any changes to the test network, this struct must be used.
- */
+*/
 type ServiceNetworkImpl struct {
 	// When the network is destroyed, all requests will fail
 	// This ensures that when the initializer tells the API container to destroy everything, the still-running
 	//  testsuite can't create more work
-	isDestroyed bool   	// VERY IMPORTANT TO CHECK AT THE START OF EVERY METHOD!
+	isDestroyed bool // VERY IMPORTANT TO CHECK AT THE START OF EVERY METHOD!
 
-	mutex *sync.Mutex	// VERY IMPORTANT TO CHECK AT THE START OF EVERY METHOD!
+	mutex *sync.Mutex // VERY IMPORTANT TO CHECK AT THE START OF EVERY METHOD!
 
 	// Whether partitioning has been enabled for this particular test
 	isPartitioningEnabled bool
@@ -68,6 +70,8 @@ type ServiceNetworkImpl struct {
 	freeIpAddrTracker *commons.FreeIpAddrTracker
 
 	dockerManager *docker_manager.DockerManager
+
+	dockerNetworkId string
 
 	enclaveDataVolume *enclave_data_volume.EnclaveDataVolume
 
@@ -78,7 +82,7 @@ type ServiceNetworkImpl struct {
 	// These are separate maps, rather than being bundled into a single containerInfo-valued map, because
 	//  they're registered at different times (rather than in one atomic operation)
 	serviceRegistrationInfo map[service_network_types.ServiceID]serviceRegistrationInfo
-	serviceRunInfo map[service_network_types.ServiceID]serviceRunInfo
+	serviceRunInfo          map[service_network_types.ServiceID]serviceRunInfo
 
 	networkingSidecars map[service_network_types.ServiceID]networking_sidecar.NetworkingSidecar
 
@@ -89,6 +93,7 @@ func NewServiceNetworkImpl(
 		isPartitioningEnabled bool,
 		freeIpAddrTracker *commons.FreeIpAddrTracker,
 		dockerManager *docker_manager.DockerManager,
+		dockerNetworkId string,
 		enclaveDataVolume *enclave_data_volume.EnclaveDataVolume,
 		userServiceLauncher *user_service_launcher.UserServiceLauncher,
 		networkingSidecarManager networking_sidecar.NetworkingSidecarManager) *ServiceNetworkImpl {
@@ -98,10 +103,11 @@ func NewServiceNetworkImpl(
 		isPartitioningEnabled: isPartitioningEnabled,
 		freeIpAddrTracker:     freeIpAddrTracker,
 		dockerManager:         dockerManager,
+		dockerNetworkId:       dockerNetworkId,
 		enclaveDataVolume:     enclaveDataVolume,
 		userServiceLauncher:   userServiceLauncher,
 		mutex:                 &sync.Mutex{},
-		topology:            partition_topology.NewPartitionTopology(
+		topology: partition_topology.NewPartitionTopology(
 			defaultPartitionId,
 			defaultPartitionConnection,
 		),
@@ -114,7 +120,7 @@ func NewServiceNetworkImpl(
 
 /*
 Completely repartitions the network, throwing away the old topology
- */
+*/
 func (network *ServiceNetworkImpl) Repartition(
 		ctx context.Context,
 		newPartitionServices map[service_network_types.PartitionID]*service_network_types.ServiceIDSet,
@@ -135,7 +141,7 @@ func (network *ServiceNetworkImpl) Repartition(
 	}
 	blocklists, err := network.topology.GetBlocklists()
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting the blocklists after repartition, meaning that " +
+		return stacktrace.Propagate(err, "An error occurred getting the blocklists after repartition, meaning that "+
 			"no partitions are actually being enforced!")
 	}
 	if err := updateIpTables(ctx, blocklists, network.serviceRegistrationInfo, network.networkingSidecars); err != nil {
@@ -174,10 +180,6 @@ func (network ServiceNetworkImpl) RegisterService(
 		)
 	}
 
-	serviceDirectory, err := network.enclaveDataVolume.NewServiceDirectory(string(serviceId))
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred creating a new service directory for service with ID '%v'", serviceId)
-	}
 	ip, err := network.freeIpAddrTracker.GetFreeIpAddr()
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting an IP for service with ID '%v'", serviceId)
@@ -191,11 +193,20 @@ func (network ServiceNetworkImpl) RegisterService(
 	}()
 	logrus.Debugf("Giving service '%v' IP '%v'", serviceId, ip.String())
 
-	registrationInfo := serviceRegistrationInfo{
+	serviceGUID := newServiceGUID(serviceId)
+
+	serviceDirectory, err := network.enclaveDataVolume.GetServiceDirectory(serviceGUID)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred creating a new service directory for service with GUID '%v'", serviceGUID)
+	}
+
+	serviceRegistrationInfo := serviceRegistrationInfo{
+		serviceGUID:      serviceGUID,
 		ipAddr:           ip,
 		serviceDirectory: serviceDirectory,
 	}
-	network.serviceRegistrationInfo[serviceId] = registrationInfo
+
+	network.serviceRegistrationInfo[serviceId] = serviceRegistrationInfo
 	shouldUndoRegistrationInfoAdd := true
 	defer func() {
 		// If an error occurs, the service ID won't be used so we need to delete it from the map
@@ -317,7 +328,7 @@ Starts a previously-registered but not-started service by creating it in a conta
 Returns:
 	Mapping of port-used-by-service -> port-on-the-Docker-host-machine where the user can make requests to the port
 		to access the port. If a used port doesn't have a host port bound, then the value will be nil.
- */
+*/
 func (network *ServiceNetworkImpl) StartService(
 		ctx context.Context,
 		serviceId service_network_types.ServiceID,
@@ -353,7 +364,7 @@ func (network *ServiceNetworkImpl) StartService(
 	if network.isPartitioningEnabled {
 		blocklists, err := network.topology.GetBlocklists()
 		if err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred getting the blocklists for updating iptables before the " +
+			return nil, stacktrace.Propagate(err, "An error occurred getting the blocklists for updating iptables before the "+
 				"node was added, which means we can't add the node because we can't partition it away properly")
 		}
 		blocklistsWithoutNewNode := map[service_network_types.ServiceID]*service_network_types.ServiceIDSet{}
@@ -364,16 +375,18 @@ func (network *ServiceNetworkImpl) StartService(
 			blocklistsWithoutNewNode[serviceInTopologyId] = servicesToBlock
 		}
 		if err := updateIpTables(ctx, blocklistsWithoutNewNode, network.serviceRegistrationInfo, network.networkingSidecars); err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred updating the iptables of all the other services " +
+			return nil, stacktrace.Propagate(err, "An error occurred updating the iptables of all the other services "+
 				"before adding the node, meaning that the node wouldn't actually start in a partition")
 		}
 	}
 
 	serviceContainerId, hostPortBindings, err := network.userServiceLauncher.Launch(
 		ctx,
-		serviceId,
+		registrationInfo.serviceGUID,
+		string(serviceId),
 		registrationInfo.ipAddr,
 		imageName,
+		network.dockerNetworkId,
 		usedPorts,
 		entrypointArgs,
 		cmdArgs,
@@ -392,7 +405,7 @@ func (network *ServiceNetworkImpl) StartService(
 	network.serviceRunInfo[serviceId] = runInfo
 
 	if network.isPartitioningEnabled {
-		sidecar, err := network.networkingSidecarManager.Add(ctx, serviceId, serviceContainerId)
+		sidecar, err := network.networkingSidecarManager.Add(ctx, registrationInfo.serviceGUID, serviceContainerId)
 		if err != nil {
 			return nil, stacktrace.Propagate(err, "An error occurred creating the networking sidecar container")
 		}
@@ -406,7 +419,7 @@ func (network *ServiceNetworkImpl) StartService(
 		//  getting the blocklists not an expensive call (see also https://github.com/kurtosis-tech/kurtosis-core/issues/123 )
 		blocklists, err := network.topology.GetBlocklists()
 		if err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred getting the blocklists to know what iptables " +
+			return nil, stacktrace.Propagate(err, "An error occurred getting the blocklists to know what iptables "+
 				"updates to apply on the new node")
 		}
 		newNodeBlocklist := blocklists[serviceId]
@@ -414,7 +427,7 @@ func (network *ServiceNetworkImpl) StartService(
 			serviceId: newNodeBlocklist,
 		}
 		if err := updateIpTables(ctx, updatesToApply, network.serviceRegistrationInfo, network.networkingSidecars); err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred applying the iptables on the new node to partition it " +
+			return nil, stacktrace.Propagate(err, "An error occurred applying the iptables on the new node to partition it "+
 				"off from other nodes")
 		}
 	}
@@ -544,7 +557,7 @@ func (network *ServiceNetworkImpl) Destroy(ctx context.Context) error {
 		}
 		joinedErrStrings := strings.Join(errorStrs, "\n\n")
 		return stacktrace.NewError(
-			"One or more error(s) occurred stopping the services in the test network " +
+			"One or more error(s) occurred stopping the services in the test network "+
 				"during service network destruction:\n%s",
 			joinedErrStrings)
 	}
@@ -558,13 +571,12 @@ func (network *ServiceNetworkImpl) GetServiceIDs() map[service_network_types.Ser
 	serviceIDs := make(map[service_network_types.ServiceID]bool, len(network.serviceRunInfo))
 
 	for key, _ := range network.serviceRunInfo {
-		if _, ok := serviceIDs[key]; !ok{
+		if _, ok := serviceIDs[key]; !ok {
 			serviceIDs[key] = true
 		}
 	}
 	return serviceIDs
 }
-
 
 // ====================================================================================================
 // 									   Private helper methods
@@ -590,7 +602,14 @@ func (network *ServiceNetworkImpl) removeServiceWithoutMutex(
 			return stacktrace.Propagate(err, "An error occurred stopping the container with ID %v", serviceContainerId)
 		}
 		delete(network.serviceRunInfo, serviceId)
-		logrus.Debugf("Successfully stopped container ID")
+		logrus.Debugf("Successfully stopped container ID '%v'", serviceContainerId)
+		logrus.Debugf("Disconnecting container ID '%v' from network ID '%v'...", serviceContainerId, network.dockerNetworkId)
+		//Disconnect the container from the network in order to free the network container's alias if a new service with same alias
+		//is loaded in the network
+		if err := network.dockerManager.DisconnectContainerFromNetwork(ctx, serviceContainerId, network.dockerNetworkId); err != nil {
+			return stacktrace.Propagate(err, "An error occurred disconnecting the container with ID %v from network with ID %v", serviceContainerId, network.dockerNetworkId)
+		}
+		logrus.Debugf("Successfully disconnected container ID '%v'", serviceContainerId)
 	}
 	network.freeIpAddrTracker.ReleaseIpAddr(registrationInfo.ipAddr)
 
@@ -615,7 +634,7 @@ func (network *ServiceNetworkImpl) removeServiceWithoutMutex(
 Updates the iptables of the services with the given IDs to match the target blocklists
 
 NOTE: This is not thread-safe, so it must be within a function that locks mutex!
- */
+*/
 func updateIpTables(
 		ctx context.Context,
 		targetBlocklists map[service_network_types.ServiceID]*service_network_types.ServiceIDSet,
@@ -628,7 +647,7 @@ func updateIpTables(
 			infoForService, found := serviceRegistrationInfo[serviceIdToBlock]
 			if !found {
 				return stacktrace.NewError(
-					"Service with ID '%v' needs to block service with ID '%v', but the latter " +
+					"Service with ID '%v' needs to block service with ID '%v', but the latter "+
 						"doesn't have service registration info (i.e. an IP) associated with it",
 					serviceId,
 					serviceIdToBlock)
@@ -652,4 +671,9 @@ func updateIpTables(
 	return nil
 }
 
-
+func newServiceGUID(serviceID service_network_types.ServiceID) service_network_types.ServiceGUID {
+	now := time.Now()
+	nowSec := now.Unix()
+	registrationTimestampStr := strconv.FormatInt(nowSec, 10)
+	return service_network_types.ServiceGUID(string(serviceID) + "_" + registrationTimestampStr)
+}