+++ core/api_container/server/service_network/user_service_launcher/user_service_launcher.go
@@ -8,10 +8,10 @@ package user_service_launcher
 import (
 	"context"
 	"github.com/docker/go-connections/nat"
+	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/service_network/service_network_types"
 	"github.com/kurtosis-tech/kurtosis/api_container/server/service_network/user_service_launcher/files_artifact_expander"
 	"github.com/kurtosis-tech/kurtosis/commons"
-	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
 	"github.com/kurtosis-tech/kurtosis/commons/object_name_providers"
 	"github.com/palantir/stacktrace"
 	"net"
@@ -31,14 +31,12 @@ type UserServiceLauncher struct {
 
 	filesArtifactExpander *files_artifact_expander.FilesArtifactExpander
 
-	dockerNetworkId string
-
 	// The name of the Docker volume containing data for the enclave
 	enclaveDataVolName string
 }
 
-func NewUserServiceLauncher(dockerManager *docker_manager.DockerManager, enclaveObjNameProvider *object_name_providers.EnclaveObjectNameProvider, freeIpAddrTracker *commons.FreeIpAddrTracker, shouldPublishPorts bool, filesArtifactExpander *files_artifact_expander.FilesArtifactExpander, dockerNetworkId string, enclaveDataVolName string) *UserServiceLauncher {
-	return &UserServiceLauncher{dockerManager: dockerManager, enclaveObjNameProvider: enclaveObjNameProvider, freeIpAddrTracker: freeIpAddrTracker, shouldPublishPorts: shouldPublishPorts, filesArtifactExpander: filesArtifactExpander, dockerNetworkId: dockerNetworkId, enclaveDataVolName: enclaveDataVolName}
+func NewUserServiceLauncher(dockerManager *docker_manager.DockerManager, enclaveObjNameProvider *object_name_providers.EnclaveObjectNameProvider, freeIpAddrTracker *commons.FreeIpAddrTracker, shouldPublishPorts bool, filesArtifactExpander *files_artifact_expander.FilesArtifactExpander, enclaveDataVolName string) *UserServiceLauncher {
+	return &UserServiceLauncher{dockerManager: dockerManager, enclaveObjNameProvider: enclaveObjNameProvider, freeIpAddrTracker: freeIpAddrTracker, shouldPublishPorts: shouldPublishPorts, filesArtifactExpander: filesArtifactExpander, enclaveDataVolName: enclaveDataVolName}
 }
 
 /**
@@ -50,9 +48,11 @@ Returns:
  */
 func (launcher UserServiceLauncher) Launch(
 		ctx context.Context,
-		serviceId service_network_types.ServiceID,
+		serviceGUID service_network_types.ServiceGUID,
+		dockerContainerAlias string,
 		ipAddr net.IP,
 		imageName string,
+		dockerNetworkId string,
 		usedPorts map[nat.Port]bool,
 		entrypointArgs []string,
 		cmdArgs []string,
@@ -69,7 +69,7 @@ func (launcher UserServiceLauncher) Launch(
 	// First expand the files artifacts into volumes, so that any errors get caught early
 	// NOTE: if users don't need to investigate the volume contents, we could keep track of the volumes we create
 	//  and delete them at the end of the test to keep things cleaner
-	artifactIdsToVolumes, err := launcher.filesArtifactExpander.ExpandArtifactsIntoVolumes(ctx, serviceId, usedArtifactIdSet)
+	artifactIdsToVolumes, err := launcher.filesArtifactExpander.ExpandArtifactsIntoVolumes(ctx, serviceGUID, usedArtifactIdSet)
 	if err != nil {
 		return "", nil, stacktrace.Propagate(err, "An error occurred expanding the requested files artifacts into volumes")
 	}
@@ -97,10 +97,10 @@ func (launcher UserServiceLauncher) Launch(
 	containerId, hostPortBindings, err := launcher.dockerManager.CreateAndStartContainer(
 		ctx,
 		imageName,
-		launcher.enclaveObjNameProvider.ForUserServiceContainer(serviceId),
-		string(serviceId),
+		launcher.enclaveObjNameProvider.ForUserServiceContainer(serviceGUID),
+		dockerContainerAlias,
 		nil,	// User services won't run in interactive mode
-		launcher.dockerNetworkId,
+		dockerNetworkId,
 		ipAddr,
 		map[docker_manager.ContainerCapability]bool{},
 		docker_manager.DefaultNetworkMode,