+++ core/commons/enclave_data_volume/enclave_data_volume.go
@@ -6,8 +6,7 @@
 package enclave_data_volume
 
 import (
-	"fmt"
-	"github.com/google/uuid"
+	"github.com/kurtosis-tech/kurtosis/api_container/server/service_network/service_network_types"
 	"github.com/palantir/stacktrace"
 	"path"
 )
@@ -53,20 +52,18 @@ func (volume EnclaveDataVolume) GetStaticFileCache() (*StaticFileCache, error) {
 	return newStaticFileCache(absoluteDirpath, relativeDirpath), nil
 }
 
-// Creates a new, unique service directory for a service with the given service ID
-func (volume EnclaveDataVolume) NewServiceDirectory(serviceId string) (*ServiceDirectory, error) {
+// Get the unique service directory for a service with the given service GUID
+func (volume EnclaveDataVolume) GetServiceDirectory(serviceGUID service_network_types.ServiceGUID) (*ServiceDirectory, error) {
 	allServicesRelativeDirpath := allServicesDirname
 	allServicesAbsoluteDirpath := path.Join(volume.absMountDirpath, allServicesDirname)
 	if err := ensureDirpathExists(allServicesAbsoluteDirpath); err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred ensuring all services dirpath '%v' exists inside the enclave data dir", allServicesAbsoluteDirpath)
 	}
 
-	uniqueId := uuid.New()
-	serviceDirname := fmt.Sprintf("%v_%v", serviceId, uniqueId.String())
-	absoluteServiceDirpath := path.Join(allServicesAbsoluteDirpath, serviceDirname)
+	absoluteServiceDirpath := path.Join(allServicesAbsoluteDirpath, string(serviceGUID))
 	if err := ensureDirpathExists(absoluteServiceDirpath); err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred ensuring service dirpath '%v' exists inside the enclave data dir", absoluteServiceDirpath)
 	}
-	relativeServiceDirpath := path.Join(allServicesRelativeDirpath, serviceDirname)
+	relativeServiceDirpath := path.Join(allServicesRelativeDirpath, string(serviceGUID))
 	return newServiceDirectory(absoluteServiceDirpath, relativeServiceDirpath), nil
 }